requirements.txt
cryptography==36.0.2
dj-database-url==2.0.0
Django==4.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-debug-toolbar==3.3.0
django-hint==0.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
execnet==1.9.0
flake8==6.0.0
gunicorn==22.0.0
python-decouple==3.6
pycodestyle==2.10.0
psycopg2-binary==2.9.6
pytz==2021.3
Pillow
six==1.16.0
whitenoise==6.6.0